+++ SymbolicSummation/Real.cpp
@@ -111,7 +111,7 @@ Expression Real::divide(Expression other) const {
 };
 String Real::print() const {
     String result;
-    if(ceilf(value) == value)
+    if((float)ceilf(value) == value)
         result = std::to_string((int)value);
     else
         result = std::to_string(value);
@@ -130,7 +130,7 @@ Expression Real::distribute(Expression other) const {
         int index = positionOfType(mulMembers, REALTYPE);
         Expression result = index >= 0? thisExpr*mulMembers[index] : thisExpr;
         newMembers.push_back(result);
-        for(int i = 0; i< mulMembers.size(); i++) {
+        for(size_t i = 0; i< mulMembers.size(); i++) {
             if(i == index)
                 continue;
             newMembers.push_back(mulMembers[i]);
@@ -141,7 +141,7 @@ Expression Real::distribute(Expression other) const {
         const Add& otherAdd= dynamic_cast<const Add&>(*other);
         ExprVector newMembers = *new ExprVector();
         ExprVector addMembers = otherAdd.getMembers();
-        for(int i = 0; i< addMembers.size(); i++) {
+        for(size_t i = 0; i< addMembers.size(); i++) {
             newMembers.push_back(distribute(addMembers[i]));
         }
         return *new Expression(new Add(newMembers));
